refactor(static): inline KVItemHelper into createKV
Remove unnecessary abstraction layer by collapsing kv-helper.ts into
create-kv.ts. The KVItemHelper type was internal-only and added
indirection without benefit - each method just delegated to a pre-created
helper map.

Changes:
- Delete kv-helper.ts, inline parseValue and all logic into createKV
- Remove KVItemHelper type from types.ts and index.ts exports
- Update spec file structure documentation
