+++ packages/epicenter/src/static/create-kv.test.ts
@@ -6,12 +6,12 @@ import { defineKV } from './define-kv.js';
 
 describe('createKV', () => {
 	test('set and get a value', () => {
-		const theme = defineKV()
-			.version(type({ mode: "'light' | 'dark'" }))
-			.migrate((v) => v);
-
 		const ydoc = new Y.Doc();
-		const kv = createKV(ydoc, { theme });
+		const kv = createKV(ydoc, {
+			theme: defineKV()
+				.version(type({ mode: "'light' | 'dark'" }))
+				.migrate((v) => v),
+		});
 
 		kv.set('theme', { mode: 'dark' });
 
@@ -23,24 +23,24 @@ describe('createKV', () => {
 	});
 
 	test('get returns not_found for unset key', () => {
-		const theme = defineKV()
-			.version(type({ mode: "'light' | 'dark'" }))
-			.migrate((v) => v);
-
 		const ydoc = new Y.Doc();
-		const kv = createKV(ydoc, { theme });
+		const kv = createKV(ydoc, {
+			theme: defineKV()
+				.version(type({ mode: "'light' | 'dark'" }))
+				.migrate((v) => v),
+		});
 
 		const result = kv.get('theme');
 		expect(result.status).toBe('not_found');
 	});
 
 	test('delete removes the value', () => {
-		const theme = defineKV()
-			.version(type({ mode: "'light' | 'dark'" }))
-			.migrate((v) => v);
-
 		const ydoc = new Y.Doc();
-		const kv = createKV(ydoc, { theme });
+		const kv = createKV(ydoc, {
+			theme: defineKV()
+				.version(type({ mode: "'light' | 'dark'" }))
+				.migrate((v) => v),
+		});
 
 		kv.set('theme', { mode: 'dark' });
 		expect(kv.get('theme').status).toBe('valid');
@@ -50,19 +50,19 @@ describe('createKV', () => {
 	});
 
 	test('migrates old data on read', () => {
-		const theme = defineKV()
-			.version(type({ mode: "'light' | 'dark'" }))
-			.version(type({ mode: "'light' | 'dark'", fontSize: 'number' }))
-			.migrate((v) => {
-				if (!('fontSize' in v)) return { ...v, fontSize: 14 };
-				return v;
-			});
-
 		const ydoc = new Y.Doc();
-		const kv = createKV(ydoc, { theme });
+		const kv = createKV(ydoc, {
+			theme: defineKV()
+				.version(type({ mode: "'light' | 'dark'" }))
+				.version(type({ mode: "'light' | 'dark'", fontSize: 'number' }))
+				.migrate((v) => {
+					if (!('fontSize' in v)) return { ...v, fontSize: 14 };
+					return v;
+				}),
+		});
 
 		// Simulate old data
-		const yarray = ydoc.getArray<{ key: string; val: unknown }>('static:kv');
+		const yarray = ydoc.getArray<{ key: string; val: unknown }>('kv');
 		yarray.push([{ key: 'theme', val: { mode: 'dark' } }]);
 
 		// Read should migrate
